+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Added extra data option for MAs to tell PG to ignore patching a shape (see wiki)
 - Fixed output exist check happening after existing output deleting
 - If MO2 is selected and not launched from MO2 a critical error will not be shown
+- Light placer JSONs will now be patched if PG created a duplicate NIF
 
 ## [0.8.12] - 2025-05-18